+++ pkg/api/study-service_grpc.pb.go
@@ -80,6 +80,8 @@ type StudyServiceApiClient interface {
 	GetStudyResponseStatistics(ctx context.Context, in *SurveyResponseQuery, opts ...grpc.CallOption) (*StudyResponseStatistics, error)
 	StreamStudyResponses(ctx context.Context, in *SurveyResponseQuery, opts ...grpc.CallOption) (StudyServiceApi_StreamStudyResponsesClient, error)
 	StreamParticipantStates(ctx context.Context, in *ParticipantStateQuery, opts ...grpc.CallOption) (StudyServiceApi_StreamParticipantStatesClient, error)
+	GetParticipantStatesWithPagination(ctx context.Context, in *GetPStatesWithPaginationQuery, opts ...grpc.CallOption) (*ParticipantStatesWithPagination, error)
+	GetParticipantStateByID(ctx context.Context, in *ParticipantStateByIDQuery, opts ...grpc.CallOption) (*ParticipantState, error)
 	StreamReportHistory(ctx context.Context, in *ReportHistoryQuery, opts ...grpc.CallOption) (StudyServiceApi_StreamReportHistoryClient, error)
 	StreamParticipantFileInfos(ctx context.Context, in *FileInfoQuery, opts ...grpc.CallOption) (StudyServiceApi_StreamParticipantFileInfosClient, error)
 	GetConfidentialResponses(ctx context.Context, in *ConfidentialResponsesQuery, opts ...grpc.CallOption) (*ConfidentialResponses, error)
@@ -624,6 +626,24 @@ func (x *studyServiceApiStreamParticipantStatesClient) Recv() (*ParticipantState
 	return m, nil
 }
 
+func (c *studyServiceApiClient) GetParticipantStatesWithPagination(ctx context.Context, in *GetPStatesWithPaginationQuery, opts ...grpc.CallOption) (*ParticipantStatesWithPagination, error) {
+	out := new(ParticipantStatesWithPagination)
+	err := c.cc.Invoke(ctx, "/influenzanet.study_service.StudyServiceApi/GetParticipantStatesWithPagination", in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *studyServiceApiClient) GetParticipantStateByID(ctx context.Context, in *ParticipantStateByIDQuery, opts ...grpc.CallOption) (*ParticipantState, error) {
+	out := new(ParticipantState)
+	err := c.cc.Invoke(ctx, "/influenzanet.study_service.StudyServiceApi/GetParticipantStateByID", in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
 func (c *studyServiceApiClient) StreamReportHistory(ctx context.Context, in *ReportHistoryQuery, opts ...grpc.CallOption) (StudyServiceApi_StreamReportHistoryClient, error) {
 	stream, err := c.cc.NewStream(ctx, &StudyServiceApi_ServiceDesc.Streams[4], "/influenzanet.study_service.StudyServiceApi/StreamReportHistory", opts...)
 	if err != nil {
@@ -894,6 +914,8 @@ type StudyServiceApiServer interface {
 	GetStudyResponseStatistics(context.Context, *SurveyResponseQuery) (*StudyResponseStatistics, error)
 	StreamStudyResponses(*SurveyResponseQuery, StudyServiceApi_StreamStudyResponsesServer) error
 	StreamParticipantStates(*ParticipantStateQuery, StudyServiceApi_StreamParticipantStatesServer) error
+	GetParticipantStatesWithPagination(context.Context, *GetPStatesWithPaginationQuery) (*ParticipantStatesWithPagination, error)
+	GetParticipantStateByID(context.Context, *ParticipantStateByIDQuery) (*ParticipantState, error)
 	StreamReportHistory(*ReportHistoryQuery, StudyServiceApi_StreamReportHistoryServer) error
 	StreamParticipantFileInfos(*FileInfoQuery, StudyServiceApi_StreamParticipantFileInfosServer) error
 	GetConfidentialResponses(context.Context, *ConfidentialResponsesQuery) (*ConfidentialResponses, error)
@@ -1052,6 +1074,12 @@ func (UnimplementedStudyServiceApiServer) StreamStudyResponses(*SurveyResponseQu
 }
 func (UnimplementedStudyServiceApiServer) StreamParticipantStates(*ParticipantStateQuery, StudyServiceApi_StreamParticipantStatesServer) error {
 	return status.Errorf(codes.Unimplemented, "method StreamParticipantStates not implemented")
+}
+func (UnimplementedStudyServiceApiServer) GetParticipantStatesWithPagination(context.Context, *GetPStatesWithPaginationQuery) (*ParticipantStatesWithPagination, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method GetParticipantStatesWithPagination not implemented")
+}
+func (UnimplementedStudyServiceApiServer) GetParticipantStateByID(context.Context, *ParticipantStateByIDQuery) (*ParticipantState, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method GetParticipantStateByID not implemented")
 }
 func (UnimplementedStudyServiceApiServer) StreamReportHistory(*ReportHistoryQuery, StudyServiceApi_StreamReportHistoryServer) error {
 	return status.Errorf(codes.Unimplemented, "method StreamReportHistory not implemented")
@@ -1971,6 +1999,42 @@ func (x *studyServiceApiStreamParticipantStatesServer) Send(m *ParticipantState)
 	return x.ServerStream.SendMsg(m)
 }
 
+func _StudyServiceApi_GetParticipantStatesWithPagination_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(GetPStatesWithPaginationQuery)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(StudyServiceApiServer).GetParticipantStatesWithPagination(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/influenzanet.study_service.StudyServiceApi/GetParticipantStatesWithPagination",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(StudyServiceApiServer).GetParticipantStatesWithPagination(ctx, req.(*GetPStatesWithPaginationQuery))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _StudyServiceApi_GetParticipantStateByID_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(ParticipantStateByIDQuery)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(StudyServiceApiServer).GetParticipantStateByID(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/influenzanet.study_service.StudyServiceApi/GetParticipantStateByID",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(StudyServiceApiServer).GetParticipantStateByID(ctx, req.(*ParticipantStateByIDQuery))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
 func _StudyServiceApi_StreamReportHistory_Handler(srv interface{}, stream grpc.ServerStream) error {
 	m := new(ReportHistoryQuery)
 	if err := stream.RecvMsg(m); err != nil {
@@ -2315,6 +2379,14 @@ var StudyServiceApi_ServiceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "GetStudyResponseStatistics",
 			Handler:    _StudyServiceApi_GetStudyResponseStatistics_Handler,
+		},
+		{
+			MethodName: "GetParticipantStatesWithPagination",
+			Handler:    _StudyServiceApi_GetParticipantStatesWithPagination_Handler,
+		},
+		{
+			MethodName: "GetParticipantStateByID",
+			Handler:    _StudyServiceApi_GetParticipantStateByID_Handler,
 		},
 		{
 			MethodName: "GetConfidentialResponses",